requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.2
django-appconf==1.0.5
django-autoslug==1.9.8
django-countries==7.4.2
django-imagekit==4.1.0
django-starcross-gallery==1.0.14
pilkit==2.0
Pillow==9.1.1
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
